Remove onRowDataChange from being passed as prop
Since onRowDataChange is not being removed as a prop being passed into child components, and it isn't used by them, this is eventually being put onto a simple div element which causes an error because it isn't recognized. By removing this property from the props that are passed down, this error goes away because ultimately the prop function 'onRowDataChange' isn't being loaded onto the subsequent div element at the bottom of the component chain. When a row is updated, the hook 'onRowDataChange' is still working as expected.