Fix rows becoming locked when edit state is lost
Edit mode is captured on the tableData field of individual row data.
This field is overwritten when new row data is rendered for the table.
Overwriting this field without dropping the the lastEditingRow in
DataManager causes the entire table to become locked.

Add call to DataManager#changeRowEditing to drop all references to
editing rows. This prevents table from becoming locked when row data is
updated while editing.

https://github.com/mbrn/material-table/issues/1610