Ensure path data is not processed multiple times
The function `addRow` calls itself to add the `dataRows` parent `addRow(parent)`. This causes a reprocessing of the parent row, depending on the order of the data data source. If the parent row now has another sibling its path will be updated with the wrong index due to `parent.tableData.childRows.length - 1`. 

The `addRow(parent)` line is needed to ensure that the parent rows `tableData` is configured. However, once configured it shouldn't need to be reprocessed and can be skipped.

I believe adding check before setting the path value will solve an issue I've seen when searching tree based data. The error occures in markForTreeRemove while it tries to walk the hierarchy using path. see issue #958