Add align to tableCells from columnDef
This enables the override of the cell align to match the available props from material-ui.
By adding align to the column def, it will be passed down to the header and the cells for that column. It will override the default align which is either left or right for numeric values.
If not provided, the previous used alignment strategy will be the fallback.
The types are updated as well with all possible options:
'center' |'inherit' | 'justify'|'left' |'right';